add map.js file and refactor logic
